backend/requirements.txt
# --- Web / API ---
Flask>=3.0,<4
flask-jwt-extended>=4.6
Flask-Cors>=4.0
Flask-Bcrypt>=1.0
Flask-PyMongo>=2.3
email_validator>=2.0

# --- CV / Audio stack ---
mediapipe==0.10.14
opencv-python==4.10.0.84

# librosa 0.10 does not support numpy 2.x -> lock to 1.x
numpy>=1.24,<2.0
scipy>=1.10,<2.0
soundfile>=0.12
librosa>=0.10,<0.11

# Use ffmpeg inside the venv (no system PATH required)
moviepy<2.0
imageio-ffmpeg>=0.4.8


# NLP tools
language-tool-python>=2.8,<3.0
# NLP tools
language-tool-python>=2.8,<3.0

# Whisper ASR (pulls torch)
openai-whisper

# --- ML / Training ---
scikit-learn>=1.3
joblib>=1.3
# Use compatible TensorFlow version for Python 3.11
tensorflow>=2.12,<2.21
pandas>=2.0

# --- AI / Gemini ---
google-generativeai>=0.3.0

# --- Testing & Development ---
pytest>=7.0
pytest-flask>=1.2
pytest-mock>=3.10
requests-mock>=1.10
colorama>=0.4.6
unittest-xml-reporting>=3.2.0

# Testing dependencies
pytest>=7.0
pytest-flask>=1.2
pytest-mock>=3.10
requests-mock>=1.10

language_tool_python>=2.7.0
openai-whisper>=20231117

scikit-learn>=1.3
joblib>=1.3
language_tool_python>=2.7.0
